+++ src/Exception/InvalidArgumentException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace DevopsToolAzureBlobFilesystemSupport\Exception;
+namespace ConductorAzureBlobFilesystemSupport\Exception;
 
 class InvalidArgumentException extends \InvalidArgumentException implements ExceptionInterface
 {